@@ -7,12 +7,13 @@
     "test": "jest"
   },
   "dependencies": {
-    "react": "15.3.2",
-    "react-native": "^0.36.1",
-    "react-native-scrollable-tab-view": "^0.6.0",
+    "react": "^16.0.0-alpha.12",
+    "react-native": "^0.45.1",
+    "react-native-deprecated-custom-components": "^0.1.0",
+    "react-native-scrollable-tab-view": "^0.7.4",
     "react-native-swiper": "^1.5.3",
     "react-native-tab-navigator": "^0.3.3",
-    "react-native-vector-icons": "^3.0.0"
+    "react-native-vector-icons": "^4.2.0"
   },
   "jest": {
     "preset": "jest-react-native"
